Daily Sales Report(DSR) with automated location tracking
We are going to develop software to manage the Daily Sales Report (DSR) on smart devices. Application provide the facility to maintain DSR on daily basis. Managing DSR manually by vendors and MR is very difficult, there are chances of fake report provided by employee which result in loss of organization.